+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.2] - 2024-11-16
 ### Changed
 - Updated the edition from 2018 to 2021.
+
+### Fixed
+- Fixed `enable_adaptive_data_clocking` and `disable_adaptive_data_clocking`
+  functions in the `mpsse!` macro.
 
 ## [0.1.1] - 2022-02-01
 ### Added
@@ -16,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release.
 
-[Unreleased]: https://github.com/ftdi-rs/ftdi-mpsse/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/ftdi-rs/ftdi-mpsse/compare/v0.1.2...HEAD
+[0.1.2]: https://github.com/ftdi-rs/ftdi-mpsse/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/ftdi-rs/ftdi-mpsse/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/ftdi-rs/ftdi-mpsse/releases/tag/v0.1.0